+++ mmros/src/tensorrt/plugin/grid_sampler/trt_grid_sampler_kernel.cu
@@ -5,6 +5,8 @@
 #include "mmros/tensorrt/plugin/grid_sampler/trt_grid_sampler.hpp"
 #include "mmros/tensorrt/plugin/grid_sampler/trt_grid_sampler_kernel.hpp"
 
+#include <cstdint>
+
 namespace mmros::plugin  // NOLINT
 {
 using mmros::plugin::TensorDesc;
@@ -349,7 +351,7 @@ __global__ void grid_sampler_3d_kernel(
   }
 }
 
-void create_desc(const int * dims, int nb_dims, TensorDesc & desc)
+void create_desc(const int64_t * dims, int nb_dims, TensorDesc & desc)
 {
   memcpy(&desc.shape[0], dims, sizeof(int) * nb_dims);
   desc.stride[nb_dims - 1] = 1;
@@ -360,9 +362,9 @@ void create_desc(const int * dims, int nb_dims, TensorDesc & desc)
 
 template <typename T>
 void grid_sample(
-  T * output, const T * input, const T * grid, int * output_dims, int * input_dims, int * grid_dims,
-  int nb_dims, GridSamplerInterpolation interp, GridSamplerPadding padding, bool align_corners,
-  cudaStream_t stream)
+  T * output, const T * input, const T * grid, int64_t * output_dims, int64_t * input_dims,
+  int64_t * grid_dims, int nb_dims, GridSamplerInterpolation interp, GridSamplerPadding padding,
+  bool align_corners, cudaStream_t stream)
 {
   TensorDesc input_desc;
   create_desc(input_dims, nb_dims, input_desc);
@@ -395,7 +397,7 @@ void grid_sample(
 }
 
 template void grid_sample<float>(
-  float * output, const float * input, const float * grid, int * output_dims, int * input_dims,
-  int * grid_dims, int nb_dims, GridSamplerInterpolation interp, GridSamplerPadding padding,
-  bool align_corners, cudaStream_t stream);
+  float * output, const float * input, const float * grid, int64_t * output_dims,
+  int64_t * input_dims, int64_t * grid_dims, int nb_dims, GridSamplerInterpolation interp,
+  GridSamplerPadding padding, bool align_corners, cudaStream_t stream);
 }  // namespace mmros::plugin
